refactor(epic-plan-decompose): thin epic-plan-decompose into phase pipeline (resolves #2495)
